Fix failing unit test, move legacy hitobjects to Objects/Legacy/.
