+++ osu.Game/Modes/Objects/Legacy/LegacySlider.cs
@@ -1,13 +1,16 @@
 ﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using osu.Game.Modes.Objects.Types;
 using System.Collections.Generic;
+using osu.Game.Modes.Objects.Types;
 using OpenTK;
 
-namespace osu.Game.Modes.Objects
+namespace osu.Game.Modes.Objects.Legacy
 {
-    internal class Slider : HitObject, IHasCurve, IHasPosition, IHasDistance, IHasRepeats, IHasCombo
+    /// <summary>
+    /// Base Slider-type, used for parsing Beatmaps.
+    /// </summary>
+    public sealed class LegacySlider : HitObject, IHasCurve, IHasPosition, IHasDistance, IHasRepeats, IHasCombo
     {
         public List<Vector2> ControlPoints { get; set; }
         public CurveType CurveType { get; set; }
